Project generator, with projects:create command
