--- lib/pakyow/generators/app/templates/%dot%env.example
@@ -1,3 +1,0 @@
-# Please place the following environment variables in your local `.env` file.
-
-SESSION_SECRET=sekret
