--- lib/pakyow/generators/app/templates/%dot%gitignore
@@ -1,8 +1,0 @@
-# ignore bundler config
-.bundle
-
-# ignore environment files
-.env*
-
-# ignore binstubs
-bin
