--- lib/pakyow/generators/app/templates/%dot%rspec
@@ -1,3 +1,0 @@
---color
---order random
---format documentation
