--- lib/pakyow/generators/app/templates/Gemfile.tt
@@ -1,18 +1,0 @@
-source 'https://rubygems.org'
-<% if generating_locally? %>
-gem 'pakyow', github: 'pakyow/pakyow'
-<% else %>
-gem 'pakyow', '~> <%= Pakyow::VERSION %>'
-<% end %>
-# app server
-gem 'puma', platforms: :ruby
-gem 'thin', platforms: :mswin
-
-# use dotenv to load environment variables
-gem 'dotenv'
-
-group :test do
-  gem 'rspec'
-end
-
-# your gems here
