--- lib/pakyow/generators/app/templates/README.md
@@ -1,29 +1,0 @@
-This is a Pakyow v1.0 project.
-
-# Getting Started
-
-Start the server:
-
-  `bundle exec pakyow server`
-
-You'll find your app running at [http://localhost:3000](http://localhost:3000)!
-
-Need to interact with your code? Fire up a console:
-
-  `bundle exec pakyow console`
-
-# Next Steps
-
-- Read the [overview](https://pakyow.org/docs/overview) to better understand the goals and architecture of Pakyow.
-- Do the [warmup](https://pakyow.org/docs/warmup) to build &amp; deploy your first project.
-- Browse the [docs](https://pakyow.org/docs) to learn more about presentation, routing, realtime channels, and more.
-- Check out the [code](https://github.com/pakyow/pakyow) on Github.
-
-We'd love to have you involved. Here are a few places to start:
-
-- [Give us a star](https://github.com/pakyow/pakyow)
-- [Participate in chat](https://gitter.im/pakyow/chat)
-- [Join the forums](http://forums.pakyow.org/)
-- [Work on a starter issue](https://github.com/pakyow/pakyow/labels/Starter)
-- [Report problems](https://github.com/pakyow/pakyow/issues)
-- [Tell your friends](https://twitter.com/share?text=Pakyow,%20build%20modern%20apps%20that%20don%27t%20break%20the%20web&hashtags=pakyow&url=https://pakyow.org)
