--- lib/pakyow/generators/app/templates/Rakefile
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-require "pakyow-rake"
-
-# put your rake task here
