--- lib/pakyow/generators/app/templates/app/lib/bindings.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-Pakyow::App.bindings do
-  # define bindings here
-end
