--- lib/pakyow/generators/app/templates/app/lib/helpers.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-module Pakyow::Helpers
-  # define methods here that are available from routes, bindings, etc
-end
