--- lib/pakyow/generators/app/templates/app/lib/routes.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-Pakyow::App.routes do
-  default do
-    logger.info "hello"
-  end
-
-  # define application routes here
-end
