--- lib/pakyow/generators/app/templates/app/views/_templates/default.html.tt
@@ -1,33 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>
-    <%= app_name %>
-  </title>
-
-  <meta content="width=device-width, initial-scale=1.0, maximum-scale=1.0" name="viewport">
-
-  <link rel="stylesheet" type="text/css" href="/styles/pakyow-css/reset.css">
-  <link rel="stylesheet" type="text/css" href="/styles/pakyow-css/structure.css">
-  <link rel="stylesheet" type="text/css" href="/styles/pakyow-css/syntax.css">
-  <link rel="stylesheet" type="text/css" href="/styles/pakyow-css/theme.css">
-
-  <script src="/scripts/ring/pakyow.js"></script>
-
-  <meta charset="UTF-8">
-</head>
-
-<body>
-  <main class="container-1">
-    <!-- @container -->
-  </main>
-
-  <footer class="container-1 margin-t">
-    <p>
-      Built with
-      <svg class="icon" width="10" height="10" viewBox="0 0 1792 1792" xmlns="http://www.w3.org/2000/svg"><path d="M896 1664q-26 0-44-18l-624-602q-10-8-27.5-26t-55.5-65.5-68-97.5-53.5-121-23.5-138q0-220 127-344t351-124q62 0 126.5 21.5t120 58 95.5 68.5 76 68q36-36 76-68t95.5-68.5 120-58 126.5-21.5q224 0 351 124t127 344q0 221-229 450l-623 600q-18 18-44 18z"></svg>
-      and <a href="http://pakyow.org">Pakyow</a>.
-    </p>
-  </footer>
-</body>
-</html>
