--- lib/pakyow/generators/app/templates/app/views/index.html
@@ -1,96 +1,0 @@
-<h1>
-  Welcome to Pakyow :)
-</h1>
-
-<p class="lead margin-t">
-  The following links might be handy:
-</p>
-
-<ul>
-  <li>
-    <a href="http://pakyow.org/">Official Website</a>
-  </li>
-
-  <li>
-    <a href="http://pakyow.org/docs">Guides &amp; Documentation</a>
-  </li>
-
-  <li>
-    <a href="http://github.com/pakyow/pakyow">Source Code</a>
-  </li>
-</ul>
-
-<h3 class="margin-t">
-  Show Your Support
-</h3>
-
-<p>
-  Give us a star on <a href="http://github.com/pakyow">Github</a>. But don't
-  stop there &ndash; tell your friends about the project! Writing blog posts,
-  giving talks to local user groups, and screencasting your experiences with
-  Pakyow all help spread the word. We'll return the favor by helping promote
-  the things you create. Also consider adding a link back to pakyow.org from your
-  project.
-</p>
-
-<ul class="plain social-links inline margin-t">
-  <li>
-    <iframe src="https://ghbtns.com/github-btn.html?user=pakyow&repo=pakyow&type=star&count=true&v=2" frameborder="0" scrolling="0" width="85px" height="20px"></iframe>
-  </li>
-
-  <li>
-    <a href="https://twitter.com/pakyow" class="twitter-follow-button" data-show-count="false">Follow @pakyow</a>
-    <script>!function(d,s,id){var js,fjs=d.getElementsByTagName(s)[0],p=/^http:/.test(d.location)?'http':'https';if(!d.getElementById(id)){js=d.createElement(s);js.id=id;js.src=p+'://platform.twitter.com/widgets.js';fjs.parentNode.insertBefore(js,fjs);}}(document, 'script', 'twitter-wjs');</script>
-  </li>
-
-  <li>
-    <div id="fb-root"></div>
-    <script>(function(d, s, id) {
-      var js, fjs = d.getElementsByTagName(s)[0];
-      if (d.getElementById(id)) return;
-      js = d.createElement(s); js.id = id;
-      js.src = "//connect.facebook.net/en_US/sdk.js#xfbml=1&version=v2.3&appId=1611480632444271";
-      fjs.parentNode.insertBefore(js, fjs);
-    }(document, 'script', 'facebook-jssdk'));</script>
-    <div class="fb-like" data-href="https://www.facebook.com/pakyow.rb" data-layout="button_count" data-action="like" data-show-faces="false" data-share="false"></div>
-  </li>
-
-  <li>
-    <!-- Place this tag in your head or just before your close body tag. -->
-    <script src="https://apis.google.com/js/platform.js" async defer></script>
-    <!-- Place this tag where you want the +1 button to render. -->
-    <div class="g-plusone" data-size="medium" data-href="http://pakyow.org"></div>
-  </li>
-</ul>
-
-<div style="clear:both"></div>
-
-<h3 class="margin-t">
-  Join the Conversation
-</h3>
-
-<p>
-  Come hang out on our <a href="https://gitter.im/pakyow/chat">Gitter channel</a>,
-  which took the place of our IRC channel in 2015. You should also join the
-  <a href="http://forums.pakyow.org">forums</a>, where early discussions about new Pakyow features often take place.
-  Don't forget to lend a hand &ndash; ask and answer <a href="http://stackoverflow.com/questions/tagged/pakyow">questions on Stack Overflow</a>.
-</p>
-
-<h3 class="margin-t">
-  Contribute to the Project
-</h3>
-
-<p>
-  We'd love your contributions to Pakyow code or documentation. Everything is hosted publicly on <a href="http://github.com/pakyow">Github</a>.
-  Looking through the list of open issues would be a great place to start. If you need more direction, jump on
-  one of the communication channels listed above; we're happy to help!
-</p>
-
-<h3 class="margin-t">
-  Report Problems you Encounter
-</h3>
-
-<p>
-  Found a bug or have an idea for a new feature?
-  Open up an issue on our <a href="https://github.com/pakyow/pakyow/issues">Github issues page</a>.
-</p>
