--- lib/pakyow/generators/app/templates/config.ru
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-require "./config/environment"
-run Pakyow.setup(env: ENV["RACK_ENV"])
