--- lib/pakyow/generators/app/templates/config/environment.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-require "bundler/setup"
-require "pakyow/integrations/bundler"
-require "pakyow/integrations/dotenv"
-require "./app/define"
-
-Pakyow.configure do
-  mount Pakyow::App, at: "/"
-end
