--- lib/pakyow/generators/app/templates/public/styles/pakyow-css/VERSION
@@ -1,1 +1,0 @@
-0.2.1
