--- lib/pakyow/generators/app/templates/public/styles/pakyow-css/theme.css
@@ -1,2 +1,0 @@
-html{font-size:62.5%;line-height:62.5%}body{margin:20px}body,input,button{font-size:13px;font-size:1.3rem;font-family:"Helvetica Neue",Helvetica,sans-serif}code,code *{font-size:13px;font-size:1.3rem;line-height:18.2px;line-height:1.82rem;margin-bottom:18.2px;margin-bottom:1.82rem;font-family:"Source Code Pro",Menlo,monospace}pre{line-height:0}p code{font-size:11.7px;font-size:1.17rem;display:inline}h1{font-size:32.5px;font-size:3.25rem;line-height:45.5px;line-height:4.55rem;margin-bottom:22.75px;margin-bottom:2.275rem;margin-top:34.125px;margin-top:3.4125rem}h2{font-size:24.7px;font-size:2.47rem;line-height:34.58px;line-height:3.458rem;margin-bottom:17.29px;margin-bottom:1.729rem;margin-top:25.935px;margin-top:2.5935rem}h3{font-size:18.2px;font-size:1.82rem;line-height:25.48px;line-height:2.548rem;margin-bottom:12.74px;margin-bottom:1.274rem;margin-top:19.11px;margin-top:1.911rem}h4{font-size:15.6px;font-size:1.56rem;line-height:21.84px;line-height:2.184rem;margin-bottom:10.92px;margin-bottom:1.092rem;margin-top:16.38px;margin-top:1.638rem}h5{font-size:13px;font-size:1.3rem;line-height:18.2px;line-height:1.82rem;margin-bottom:9.1px;margin-bottom:0.91rem;margin-top:13.65px;margin-top:1.365rem}p{font-size:13px;font-size:1.3rem;line-height:18.2px;line-height:1.82rem;margin-bottom:9.1px;margin-bottom:0.91rem}strong{font-weight:bold}em{font-style:italic}hr{margin-top:18.2px;margin-top:1.82rem;margin-bottom:18.2px;margin-bottom:1.82rem}.lead,.lead *{font-size:15.6px;font-size:1.56rem;line-height:21.84px;line-height:2.184rem;margin-bottom:21.84px;margin-bottom:2.184rem}.lead code,.lead * code{font-size:14.04px;font-size:1.404rem}ul,ol{margin-bottom:9px;margin-bottom:0.9rem;margin-left:25px;margin-left:2.5rem}ul{list-style:disc}ol{list-style:decimal}li{font-size:13px;font-size:1.3rem;line-height:18.2px;line-height:1.82rem}ul.horizontal,ol.horizontal{margin:0;display:inline}ul.horizontal li,ol.horizontal li{display:inline}ul.plain,ol.plain{margin-left:0}ul.plain li,ol.plain li{list-style:none}label{display:block}label.inline{display:inline}fieldset{margin-bottom:18.2px;margin-bottom:1.82rem}fieldset ul{margin-left:0;list-style:none}fieldset ul li{margin-bottom:9.1px;margin-bottom:0.91rem}fieldset ul li label{margin-bottom:9.1px;margin-bottom:0.91rem}input,textarea{outline:none}body{color:#222}h1,h2,h3,h4,h5{font-weight:600}h1{border-bottom:1px solid #eee;padding-bottom:5.6875px;padding-bottom:0.56875rem}code{display:block;background:#fff;padding:10px 15px;border-radius:2px;border:1px solid #eee;overflow:auto}p code{padding:1px 2px;background:#edf4fa;color:#2f6f9f;border:0}hr{border:0;height:1px;background:#eee}a{color:#e74c3c}a:hover{text-decoration:none}label{font-weight:600}footer{border-top:1px solid #eee;padding-top:13.65px;padding-top:1.365rem}.lead{color:#555}.error{font-weight:normal;background:#e74c3c;color:#fff;padding:2px 5px;border-radius:2px}.icon{fill:#e74c3c}input,textarea{padding:3px}input[type='text'],input[type='password'],textarea{border-radius:2px;border:1px solid #ccc;padding:5px}input[type='text']:focus,input[type='password']:focus,textarea:focus{border-color:#2f6f9f}input[type='submit'],input[type='button'],button,.button{cursor:pointer;border-radius:2px;border:1px solid #ccc;padding:5px 8px;background:#ffffff;background:-moz-linear-gradient(top, #fff 0%, #f3f3f3 100%);background:-webkit-gradient(linear, left top, left bottom, color-stop(0%, #fff), color-stop(100%, #f3f3f3));background:-webkit-linear-gradient(top, #fff 0%, #f3f3f3 100%);background:-o-linear-gradient(top, #fff 0%, #f3f3f3 100%);background:-ms-linear-gradient(top, #fff 0%, #f3f3f3 100%);background:linear-gradient(top, #fff 0%, #f3f3f3 100%);filter:progid:DXImageTransform.Microsoft.gradient( startColorstr='#ffffff', endColorstr='#f3f3f3',GradientType=0 );color:#222;font-weight:600}input[type='submit']:hover,input[type='button']:hover,button:hover,.button:hover{text-decoration:none;background:-webkit-linear-gradient(top, #fff 0%, #f6f6f6 100%);border-color:#ccc;color:#777}
-/*# sourceMappingURL=theme.css.map */
