--- lib/pakyow/generators/app/templates/spec/spec_helper.rb
@@ -1,7 +1,0 @@
-require 'pakyow-test'
-
-Pakyow::TestHelp.setup
-
-RSpec.configure do |config|
-  config.include Pakyow::TestHelp::Helpers
-end
