+++ lib/pakyow/generators/project/default/%dot%env.erb
