+++ lib/pakyow/generators/project/default/%dot%gitignore
@@ -1,0 +1,11 @@
+# Ignore bundler config.
+#
+.bundle
+
+# Ignore environment files.
+#
+.env*
+
+# Ignore local databases.
+#
+database/*.db*
