+++ lib/pakyow/generators/project/default/%dot%ruby-version.erb
@@ -1,0 +1,1 @@
+<%= RUBY_VERSION %>
