+++ lib/pakyow/generators/project/default/Procfile
@@ -1,0 +1,1 @@
+web: bundle exec pakyow boot
