+++ lib/pakyow/generators/project/default/README.md
@@ -1,0 +1,14 @@
+# Hello, Web
+
+Start your project:
+
+  * From a command prompt, run `pakyow boot`
+
+You can access your project at [`localhost:3000`](http://localhost:3000) from your web browser.
+
+## Learn more
+
+  * Website: https://pakyow.com
+  * Community: https://pakyow.com/talk
+  * Documentation: https://pakyow.com/docs/hello
+  * View Source: https://github.com/pakyow/pakyow
