+++ lib/pakyow/generators/project/default/config/application.rb.erb
@@ -1,0 +1,13 @@
+Pakyow.app :<%= @project_name %> do
+  # Learn more about configuring an application:
+  #
+  #   * https://pakyow.com/docs/configuration
+  #
+  configure do
+    # Application config for all environments.
+  end
+
+  configure :production do
+    # Application config for the production environment.
+  end
+end
