+++ lib/pakyow/generators/project/default/frontend/layouts/default.html.erb
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <title>
+    <%= @project_name.capitalize %>
+  </title>
+
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+
+  <link rel="shortcut icon" href="/favicon.ico">
+</head>
+
+<body>
+  <main>
+    <!-- @container -->
+  </main>
+</body>
+</html>
