+++ lib/pakyow/generators/project/default/public/robots.txt
@@ -1,0 +1,1 @@
+Allow: /
